Make HMAC SHA256 generation use Bouncy Castle API
The `javax.crypto` API that was used to generate the SHA-256 HMAC is not
available without the Java Crypto Extensions.
This change replaces those calls with the Bouncy Castle alternatives,
which makes the code runnable on OpenJDK, which doesn't include the JCE
by default.

Also snuk in a fix for valueless query parameters.